Bug#30639503 post-push fix, remove too strict assert.
Patch for above bug added an assert intended to check that number of
fragments to start a scan on, was at least >= than the number of
root fragments 'owned' by this SPJ block.

However, if the table to be scanned is partitioned, we may find that
there are no fragments scas to be sent to a particular fragment. In such
cases 'noOfFrags' is decremented, and we end up violating the assert.

This patch just removes this assert being too strict

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>

Change-Id: Ia8ebee5de7344ddd51c8bbf118f1d5646d0b767c
(cherry picked from commit da90ef86edd0c325c73a2e0f8100b6de615fed5e)