Bug #30900988: REMOVE OPERATOR OVERLOADING FROM ORDER_WITH_SRC
ORDER_with_src has a user-defined operator ORDER* and operator->.
The comments claim this is "safe" because the code happens never to compare
ORDER_with_src (or at least did not at the time it was written).
Yet, user-defined operators are against the style guide, and make
for very confusing code. There is no good reason to keep this around.

Change-Id: I41f76f97e6735cf0f5956329d5186747e4c0eeb0