WL#6059 Make RIGHT JOIN be fully converted to LEFT JOIN
During parsing, the PT_joined_table object is created.

Post-WL, in its contextualize_tabs(), if the join type is JTT_RIGHT,
we swap tab1_node and tab2_node before contextualizing them. Thus,
T1 RIGHT JOIN T2
becomes exactly like
T2 LEFT JOIN T1,
as far as the lists "next_local", "next_name_resolution_table" and
"SELECT_LEX::join_list" are concerned (insisting: post-WL these lists
enumerate tables in the same order as if the LEFT JOIN had been
typed).

This is different from the pre-WL situation where the swapping
happened right after contextualize_tabs, and was done only on
"join_list": thus, "next_local" and "next_name_resolution_table" were
as in the original query, while "SELECT_LEX::join_list" was as in the
converted query.

Additionally, post-WL the PT_joined_table object gets its flag
JTT_RIGHT changed to JTT_LEFT at contextualize_tabs; and we record, in
the right-side argument of the now-LEFT JOIN, that it had a swapping
(tr2->join_order_swapped=true), and that the SELECT_LEX had a RIGHT
JOIN (more on this below).

The
enum Outer_join_type {
  JOIN_TYPE_INNER = 0,
  JOIN_TYPE_LEFT = 1,
  JOIN_TYPE_RIGHT = 2
};
is replaced with a bool. In essence it means JOIN_TYPE_RIGHT is gone.

Because "next_name_resolution" and "join_list" are in consistent order now,
TABLE_LIST::first_leaf_for_name_resolution() and
TABLE_LIST::last_leaf_for_name_resolution() don't need special-casing
for RIGHT JOIN and are thus simplified.
Likewise, store_top_level_join_columns() is simplified.

propagate_table_maps() and check_right_lateral_join() are removed.
The former was a helper for the latter. The latter was necessary to
handle this:
select * from t0, t1 right join lateral (select t1.a) as dt;
Consider first this:
select * from t0, lateral (select t1.a) as dt left join t1;
As LATERAL "looks to the left only", it failed in usual name resolution code, saying t1.a is unknown. Which is correct.
Back to:
select * from t0, t1 right join lateral (select t1.a) as dt;
Pre-WL we found t1.a as t1 is to the left _and_ the name resolution
code used  next_name_resolution_table list which is t0-t1-dt; still
this was an invalid query as it cannot be executed (it would require
to read t1 first, even though it's on the NULL-complementable side of
the join). So, pre-WL, such query needed a special detection, in
check_right_lateral_join(). Post-WL next_name_resolution_table list is
t0-dt-t1 and thus the query fails in usual name resolution code,
saying t1.a is unknown. Just like the LEFT JOIN variant.
So check_right_lateral_join() is removed.

SELECT_LEX::convert_right_join() is removed, as the conversion is done
differently now.

Some RAPID code is also simplified (slightly: doesn't have to test for
JOIN_TYPE_LEFT and JOIN_TYPE_RIGHT).

Bug#30887665 (Sakila) ASSERT: `SELECT_LEX->IS_RECURSIVE()' IN
JOIN::MAKE_JOIN_PLAN() is fixed by this WL, as it was a bug of
check_right_lateral_join() not flagging a wrong query. A test is added
in derived_correlated.test.

That was for the good side of the WL: scattered code and complexity is
removed.

There is one place, though, where additional complexity comes in: it's
insert_fields(). That function's job is to do, in the SELECT list, the
expansion  of 'table_name.*' (qualified asterisk) or '*'
(unqualified).
First, here is some background.

  SELECT * FROM T1 RIGHT JOIN T2 ON <cond>;
  is converted, during contextualization, to:
  SELECT * FROM T2 LEFT JOIN T1 ON <cond>;
  however the former has to return columns of T1 then of T2,
  while the latter has to return T2's then T1's.
  Post-WL, the conversion is complete: the lists 'next_local',
  'next_name_resolution_table' and SELECT_LEX::join_list are as if the user
  had typed the second query.

Now to the problem: pre-WL, insert_fields() would always find the
user-input order in next_local or next_name_resolution_table; post-WL,
if there was a RIGHT JOIN, the user-input order is nowhere readily
available.
Thus, post-WL, if:
- the user-input order is necessary in insert_fields() i.e. it's
expanding an unqualified '*' (for qualified 'table.*' it does matching
by table name so doesn't care for order of tables)
- and there was any RIGHT JOIN in the query block
- then we have to calculate the user-input order, by going through
join_list (diving down in join nests), and for that we need to know
that a certain join nest isn't in the user-input order. That's why in
contextualize_tabs() we took a note of the reversal.

So we are still keeping a trace that it was a RIGHT JOIN, but the
intention is that this trace should serve *only* in
insert_fields(). And if it stays so, the WL is a net gain, as the
complexity has been removed from several places, has  been confined to
insert_fields(), and new code should be able to never have to care
about right joins.

Here are some notes about MTR result changes.

rapid.bushy_join: three queries are not offloaded anymore. I verified that
in trunk, the same is reproducible: if, in trunk, I take each of these
queries, which have RIGHT JOIN, and rewrite them to LEFT JOIN manually,
then they are not offloaded anymore.

rapid.(cp_)with_non_recursive_bugs : EXPLAIN doesn't print
"low_priority" flag; it's a bug in
SELECT_LEX::print_delete/update_options, which look only at the first
table's lock type; so if the first table isn't updated, it has a READ
lock type and we don't print the word.
Happens in trunk too:
CREATE TABLE t1(a INT);
CREATE TABLE t2(a INT);
explain delete low_priority t1.*,t2.* from t2 left join t1 on 1;
explain delete low_priority t1.* from t2 left join t1 on 1;
explain delete low_priority t1.* from t2 right join t1 on 1;
No need to fix it, as the printout isn't critical (it's not like
when printing a view's body to the DD: so for a SELECT a missing
printout could be a real bug, but DELETE/UPDATE isn't any view's
body).

In the transformed query printed by EXPLAIN, select# changes, for the
subquery in "t1 rj (derived using subq)", as now we contextualize the
latter first (we used to contextualize both then swap; now we swap
then contextualize): the subquery may be created a little earlier or
later than before, so get a different select#.

Many RQG queries have pieces like
"t1 rj t2 on ... where t1.col=..."
so we have conversion to inner join; but the order of tables which
the planner uses as initial order is that of table_list,
which is now t2,t1. So, for cases where the cost of t2,t1 and t1,t2
are the same, or when optimizer_prune_level kicks in, we get different
plans, with tables swapped.

In the transformed query printed by EXPLAIN, sometimes the database's
name disappears or appears. This was already inconsistent, see the
first diff hunk in subquery_all.result.

LATERAL-specific error messages are replaced with "Unknown column", as
the LATERAL-specific check check_right_lateral_join() is gone.

In subquery_none.result, a "range checked for each record" disappears;
it's also the case in trunk, that if you write this query as RIGHT
JOIN you get "range checked etc" but not if LEFT JOIN; as the WL
converts to LEFT JOIN we see this effect.

In json.json_table:
SELECT * FROM JSON_TABLE(t1.jc, '$' COLUMNS
(id FOR ORDINALITY)) as jt RIGHT JOIN t1 ON t1.jc=jt.id;
now succeeds, which is correct.

Reviewed-by: Gleb Shchepa <gleb.shchepa@oracle.com>