Bug#30876990 NDB 8.0 : SCHEMA TABLE UPGRADE OCCURS BEFORE ALL MYSQLDS ARE UPGRADED
The ndb_schema table definition has been updated in the recent 8.0
versions of MySQL Cluster. When the cluster is being upgraded from a
lower version, the first MySQL Server that gets upgraded, upgrades the
ndb_schema table definition. The ndb_schema table upgrade is done by
dropping the older table definition and recreating it with the newer
definition. The older unupgraded MySQL Servers, connected to the cluster,
receive this ndb_schema table drop event and enter the read-only mode.
They become writable only after they are upgraded.

To prevent all the older MySQL Servers from going into a read only mode,
this patch changes the upgrade behaviour of 8.0 MySQL Servers to upgrade
the ndb_schema table definition only if all the API nodes connected to
the cluster are atleast running a 8.0 GA version, which has the updated
ndb_schema table definition. So during an upgrade to this version, any
MySQL Server that is being upgraded will not upgrade the ndb_schema table
if there is atleast one API node with a lower version connected to the
cluster. Any older version MySQL Servers connected to the cluster will
be writable through the whole upgrade process.

This fix uses the minimum API version value, maintained in the NDPAPI
(introduced by the fix for Bug#31016905), to decide if the ndb_schema
table can be upgraded. Due to this, the upgrade of ndb_schema table is
also withheld, when atleast one data node still connected to the cluster,
runs a version that has not been upgraded and does not have that fix yet.

Change-Id: Id133e99e2379c6f296774b7684670cd511a086c1