Bug#30202643 [INNODB] ASSERTION FAILURE: ARCH0PAGE.CC:1183:FALSE
Issue
=====
The bug exposes two issues.

a) In case the flushing of blocks by the page archiver does not succeed because
of some failure related to file operation then the main thread, if it's waiting
on the flush archiver, ends up waiting forever.
b) Although a) is for the general case, the bug was identified because of the
usage of innodb_read_only mode after deleting the archiving folder. Since the
page archiver cannot function in the innodb_read_only mode it does not make
sense for the archiver to be enabled. So this needs to be handled.

Fix
===
a) Handle this issue by not waiting for a flush with arch_mutex acquired. Move
the wait to the end of the Arch_Page_Sys::start function, after we're done
using the arch_mutex, and wait for the reset block flush holding the
arch_oper_mutex to not allow multiple clients to initiate start/stop tracking
at the same time.
b) For durable tracking return at the handler layer if the read only mode is
set. And for non-durable tracking like clone return immediately at the
archiver interface level.

RB: 23616
Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>