Bug#30901170 ROUTER FAILING TESTCASES SOMETIMES DO NOT PRINT USEFULL INFO.
Sometimes when the Router component test fails on some ASSERT_ or EXPECT_
macro there is no much information about the current state of all the processes
launched by the test.

This patch introduces a mechanism that would always dump the console output
and log content for each of the processes spawned by the component test.
This way the test creator does not need to bother chaining those to each of
the EXPECT_, ASSERT_ macros.

It also removes the existing 'manual' calls to the get_full_logfile and get_full_output
as this is now always done in a generic way.

RB: 23873