WL#13706: Increase default group_replication_autorejoin_tries
Change the default group_replication_autorejoin_tries from 0 to 3.

The default value of group_replication_autorejoin_tries is currently set
to 0, which means it is disabled. With this setting, a member that is
expelled due to network partition would not rejoin automatically.

The goal is to provide 'automatic network partition handling', including
recovering from network partitions, which is most effectively achieved
by setting group_replication_autorejoin_tries > 0

RB: 23623
Reviewed-by : Nuno Carvalho <nuno.carvalho@oracle.com>
Reviewed-by : Pedro A. Ribeiro <pedro.a.ribeiro@oracle.com>