WL#6057 Make semijoin work with single-table UPDATE/DELETE [patch 6/9, add semi-consistent read]
Add semi-consistent-read support to multi-table UPDATE.
Single-table already had.
We add "update t1 straight_join t2 set a=11 where a=6"
to innodb.innodb-semi-consistent; without the code patch,
this would time out because it would want to lock
all rows. Now it skips the non-matching, locked-by-others rows.
Removed the implementations of semi-consistent-read in TempTable engine
(which were just DBUG_ASSERT(false), preventing any call):
- it's simpler to just let the default implementations run,
which do nothing.
- as "main_table" may sometimes be a table that multi-UPDATE reads,
it may be a TempTable one, so we don't want to crash, we want a no-op
in this case.

Change-Id: I44ac4fd3328b2747447987026457f35fdc8faa18