Bug#29776083: EXPLICIT BINLOG INJECTION WITH SELF LOGGING ENGINES IS BROKEN
Problem:
========
When multiple storage engines are involved writing to the binlog
with at least one self-logging engine (ndbcluster), it should
fail with error ER_BINLOG_MULTIPLE_ENGINES_AND_SELF_LOGGING_ENGINE,
but that is not happening as binlog is ignoring statements that uses
a table that is not replicated. It doesn't check if the engine is
self-logging engine.

Solution:
=========
Binlog should ignore only if the statement uses a table that is
not replicated and doesn't have its own method of binary logging.

RB: 23600