BUG#29127302 - X PLUGINS DELAYED INITIALIZATION IS NOT NEED ANYMORE
Description
===========
X Plugin needs access to system-variables, still the "SQL service" is not
available when the plugin-install callback is executed. Because of that the
plugin-install mocks the "initialization" saying that its done, and a thread
is started(parallel) which does the initialization.

Fix
===
Plugin can access the "system-variables" using "component_sys_var" that
is available right away, thus the delayed initialization is not needed.

RB: 21439
Reviewed by: Grzegorz Szwarc <grzegorz.szwarc@oracle.com>