WL#6057 Make semijoin work with single-table UPDATE/DELETE [patch 9/9, improve safe_update_on_fly]
Documentation and corrections to the logic in safe_update_on_fly()
(these problems exist in trunk too):
- the requirement that the updates to the first table shouldn't influence
the selection of rows in next tables was enforced only if doing a table
scan over the first table. That doesn't make sense. Enforcing it
for all access paths.
- the range scan was managed under the JT_ALL label; historically, indeed
range scan was marked as JT_ALL (i.e. like a table scan), but that changed
a few years, ago, now range scan is marked with JT_RANGE, in
JOIN::get_best_combination():
      if (!pos->key) {
        if (tab->quick())
          tab->set_type(calc_join_type(tab->quick()->get_type()));
        else
          tab->set_type(JT_ALL);
so I'm moving the range scan mgmt under JT_RANGE. JT_INDEX_MERGE is
similar (uses a QUICK_SELECT too), so added a label for it.
- added an assertion to verify that JT_ALL never has a QUICK_SELECT.
- documented the complete logic of the function to the best of my
understanding.

So, compared to before, we are:
- more restrictive if the selection of rows in next tables is influenced
- less restrictice for range scan and index merge (they used to go
to the default label and then "return false"): now they can sometimes
do updates on the fly.

In partition_locking.result we see an effect of the first item above:
t2 is accessed with eq_ref, then t1 is accessed with a condition like
t1.b=t2.b; but if t2 is updated on the fly, t2.b changes;
so, when we access a 2nd row of t1, that row may not match because t2.b
has changed; whereas if we had delayed the update of t2.b, that row
would have matched. With this patch, we use a tmp table for both t2
and t1, so we have bigger Handler_% numbers.

We also remove some use of main_table::tmp_set in
Query_result_update::prepare(): code inspection suggests that it was
useless (exactly equivalent to simply letting
tr->check_option->fix_fields() fill map bits of def_read_set).
Its only side-effect was cleaning up tmp_set, and that is moved closer
to when tmp_set is really used.

Change-Id: I9a398acf64023cfcf1bed913cb14abf186f37ae6