Bug#30972758 MYSQL SERVER MIGHT CRASH IF UPGRADED BEFORE DATA NODES
After WL#11563, the ndbinfo table definitions are now retrieved from the
data nodes and then installed in the Data Dictionary. If a MySQL Server
is upgraded before the data nodes are upgraded and if the newer version
has new ndbinfo table definitions, they will not be installed in the
MySQL Server during upgrade as their defintion will not be in the data
nodes yet. Any attempt to install any views on these new tables will
subsequently fail as the tables won't exist in ndbinfo yet and ends up
crashing the MySQL Server during upgrade.

This patch fixes that by temporarily ignoring the error that occurs
during the view creation, if the data nodes are not upgraded yet and
thus allowing the upgrade to continue further. The new ndbinfo tables
and views will be available after all the data nodes are upgraded to the
newer version.

Change-Id: I56af0c38e20f26239a0c52f30b7efff360a2be5e