Bug#30766089 8.0.17: SEVERE REGRESSION IN PERFORMANCE OF SELECT COUNT(*) FROM LARGE TABLE WITH LOB
PB2-fix: Fix the test failure in test main.kill_debug in PB2.

The issue was that a debug statement set by the test wasn't being executed
because we'd disabled the read ahead code and the debug statement was executed
here. It has now been moved to the main thread.

Approved-by: Sunny Bains <sunny.bains@oracle.com>