Bug#30908328: CREATE UNDO TABLESPACE IS DELETING EXISTING UNDO TABLESPACE FILES
When a basename is used for undo datafile name in a CREATE statement,
the comparison that looks for an existing file is using the datadir as the relative
location instead of innodb-undo-directory. But if the existing file was created
also as a basename, MySQL will see that conflict before sending the command
to InnoDB. So this problem occurs when
1. There is an innodb-undo-directory other than the datadir.
2. The conflicting datafile name is a full path when the existing name was
created as a basename, or vise versa.

The fix is to attach the undo space basename to the undo directory if the undo
directory is different from the datadir when checking if the file already exists.

Approved by Rahul Agarkar in RB#24004