WL#7636 Post-push patch, enforce nested outer-joins ancestor dependencies.
During query optimization, the optimizer may use the equality set to
'move up' a joined table such that it is not depending on any tables
in its own join nest (inner joined), or on tables from its upper-nest
(outer joined). One such case may be the generic join structure:_
(ij -> inner joined, oj -> outer joined)

    t1 ij t2 oj (t3 ij t4)

This will usually be represented by a left-deep query tree structure, such as:

                 t1
                 | (ij)
                 t2
                / (oj)
             (t3
                \ (ij)
                 t4)

However, , assume that we have these join conditions between the tables

  - t1 ij t2:  t2.b = t1.a
  - t2 oj t3:  t3.c = t2.b
  - t3 ij t4:  t4.x = t3.c

This implies that t4.x has the equality set: t4.x = [t1.a,t2.b,t3.c].
The compiler may choose to use this to change the t4 join condition
to t4.x = t1.a, thus the query may be represented by this query tree instead.:

                t1
         (ij)  /  \ (oj)
              t2  t4)
         (oj) |
             (t3

This breaks an implementation limitation in ::prepareResultSet() which require
an outer-joined treeNode to be either a child or a sibling of the 'first-inner'
of the join nest it is embedded within (which is t3 for t4). The above topology
would break this requirement.

In order to handle this, the handler interface for doing join pushdowns
has the ability to detect and set the required 'm_ancestors' for each pushed tables.
The required m_ancestors are maintained in the 'first_inner' of each join nest,
and are interpreted as required ancestors for all tables in the same nest.
Thus, for the above case we would have t3.m_ancestors = [t2]. which was
later used when calculating the 'depend_parents' for t4 to also
include [t2]. That would result in the (correct) query tree below:

               t1
         (ij)  /
              t2
       (oj)  /  \  (oj)
           (t3  t4)

Which can be handled by ::prepareResultSet (t4 became a sibling of its first inner t3)

So far all was good, HOWEVER.

Current code overlooked the very much similar query:

     t1 ij t2 oj (t3 oj (t4))        - t3, t4 are now outer joined

t4 now became the 'first-inner' of it own join-nest, with t3 as its 'first-upper'.
As t4 is now the 'first-inner', it is embedded within the upper-nest having
t3 as its first inner.

The SPJ handler interface failed to handle the above case when setting up
the require 'depend_parents' for such a query. Thus resulting in unhandled
query tree dependencies being set up, and incorrect results being returned.

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>