Bug #30916395 ACCESS DENIED FOR LOCALSERVICE ON LOG FILE MYSQLROUTER.LOG
When running the Router with --install-service parameter the Router is trying
to grant the r/w rights to the log file to the LocalService user.
That leads to the problem when this log file already exists and was created by
the service itself. Then the current user does not have the rights to grant the
rights and the --install-service fails even though the proper rights for the
logfile are already in place.

This patch removes the code from --install-service that was trying to grant the
RW rights to the LocalService. If the logfile exists for some reason, and the
LocalService does not have the proper rights let's leave it to the user to fix
that (we error out in such case).

RB: 23919