Bug#31005815 MTR: ALLOCATE PORTS FOR THE MYSQLD ADMIN INTERFACE
The mysql server allows a separate administrative connection channel
to privileged administrators( having SERVICE_CONNECTION_ADMIN) where
in the max_connections limit doesn't apply.

In MTR, there is no port reserved for the admin interface. For tests
which use the admin channel, a workaround to use the very next port
which mysqld uses is currently being employed. This works for tests
which use a single mysqld server but is not ideal and there is a
need to support this mechanism within MTR itself.

The patch adds a dedicated port for each mysqld server which can
be used for the admin interface and updates tests which have the
workaround to use the allocated port. Please note that this change
only sets --admin-port but not --admin-address so that the admin
channel is not setup by default and doesn't affect the test suite.
Any test which requires the admin interface can set the option
--admin-address to enable the admin interface.

Change-Id: I1b9ece250d2fd28531c61f69dbb889d9fe14447f