Bug#31024838 process-manager fails to report PID of fail managed process
Problem
-------

The component-test's process manager waits for processes that it
started as part of the component test to terminate.

In case a process doesn't finish without a timeout, 'check_for_exit()'
throws and terminates the test with:

   C++ exception throw in ... Teardown()

The errormsg doesn't include any information which of the several
processes that were started, actually failed.

Change
------

- add the failed process PID to the failure message
- trigger a "dump of all logs" in case of wait-for-exit timed out

RB: 24066