WL#13851 Transform scalar subqueries to inner derived tables first increment (primary engines)
This patch makes transformation of scalar subqueries to derived tables
optionally available also for primary engines. It was made available
for secondary engines in WL#12885. In this WL, we introduce an
optimizer flag to enable the transformations for InnoDB and other
primary engines also. Note that for secondary engines, the
transformation will always be done, notwithstanding the optimizer
flag.

Reviewed by: roy.lyseng@oracle.com
Change-Id: I8bd9ef904c149a55f54f6febe3e93036dc577b4f