Bug #30664660 GCOLS, CRASH AND CORRUPTIONS WITH DELETE IGNORE + JSON
A JSON value which can not be meaningfully converted to one of temporal types,
could cause an assertion failure during DELETE, in a scenario, where there is
a secondary index on a virtual temporal column generated from a json column:

CREATE TABLE t (
  a JSON NOT NULL,
  b TIMESTAMP GENERATED ALWAYS AS (a) VIRTUAL NOT NULL,
  KEY (b)
) ENGINE=INNODB;
INSERT IGNORE INTO t(a) VALUES ('{}');
DELETE IGNORE FROM t;

In order to delete a secondary index entry, we have to compute the value of
the indexed column, to know what entry to search for deletion.
A value such as {} leads to conversion failure, which can be suppresed by
usage of IGNORE keyword.
In case of INSERT IGNORE, it generated a zero timestamp (0000-00-00 00:00:00).
In case of DELETE IGNORE, it lead to assertion failure, because the zeros were
not stored by MySQL to the buffer provided by InnoDB, as part of error handling,
and thus InnoDB was searching the secondary index for a random value.

Reviewed-by: Catalin Besleaga <catalin.besleaga@oracle.com>
RB:23928